+++ gdal/src/test/scala/geotrellis/contrib/vlm/gdal/GDALPathSpec.scala
@@ -19,7 +19,7 @@ package geotrellis.contrib.vlm.gdal
 import org.scalatest._
 
 
-class VSIPathSpec extends FunSpec with Matchers {
+class GDALDataPathSpec extends FunSpec with Matchers {
   val fileName = "file-1.tiff"
 
   describe("Formatting the given uris") {
@@ -29,7 +29,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val url = s"http://$filePath"
         val expectedPath = s"/vsicurl/$url"
 
-        VSIPath(url).vsiPath should be (expectedPath)
+        GDALDataPath(url).formattedPath should be (expectedPath)
       }
 
       it("http that points to gzip url") {
@@ -37,7 +37,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val url = s"http://$filePath"
         val expectedPath = s"/vsigzip//vsicurl/$url"
 
-        VSIPath(url).vsiPath should be (expectedPath)
+        GDALDataPath(url).formattedPath should be (expectedPath)
       }
 
       it("http that points to gzip with ! url") {
@@ -45,7 +45,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val url = s"http://$filePath!$fileName"
         val expectedPath = s"/vsigzip//vsicurl/http://$filePath/$fileName"
 
-        VSIPath(url).vsiPath should be (expectedPath)
+        GDALDataPath(url).formattedPath should be (expectedPath)
       }
 
       it("http that points to gz url") {
@@ -53,7 +53,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val url = s"http://$filePath"
         val expectedPath = s"/vsigzip//vsicurl/$url"
 
-        VSIPath(url).vsiPath should be (expectedPath)
+        GDALDataPath(url).formattedPath should be (expectedPath)
       }
 
       it("http that points to gz with ! url") {
@@ -61,7 +61,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val url = s"http://$filePath!$fileName"
         val expectedPath = s"/vsigzip//vsicurl/http://$filePath/$fileName"
 
-        VSIPath(url).vsiPath should be (expectedPath)
+        GDALDataPath(url).formattedPath should be (expectedPath)
       }
 
       it("zip+http url") {
@@ -69,7 +69,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val url = s"zip+http://$filePath"
         val expectedPath = s"/vsizip//vsicurl/$url"
 
-        VSIPath(url).vsiPath should be (expectedPath)
+        GDALDataPath(url).formattedPath should be (expectedPath)
       }
 
       it("zip+http with ! url") {
@@ -77,7 +77,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val url = s"zip+http://$filePath!$fileName"
         val expectedPath = s"/vsizip//vsicurl/http://$filePath/$fileName"
 
-        VSIPath(url).vsiPath should be (expectedPath)
+        GDALDataPath(url).formattedPath should be (expectedPath)
       }
     }
 
@@ -87,7 +87,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"file://$filePath"
         val expectedPath = filePath
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("file that points to zip uri") {
@@ -95,7 +95,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"file://$filePath"
         val expectedPath = s"/vsizip/$filePath"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("file that points to zip with ! uri") {
@@ -103,7 +103,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"file://$filePath!$fileName"
         val expectedPath = s"/vsizip/$filePath/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("zip+file uri") {
@@ -111,7 +111,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"zip+file://$path"
         val expectedPath = "/vsizip//tmp/some/data/data.zip"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("zip+file with ! uri") {
@@ -119,7 +119,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"zip+file://$path!$fileName"
         val expectedPath = s"/vsizip/$path/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
     }
 
@@ -129,7 +129,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"s3://$filePath"
         val expectedPath = s"/vsis3/$filePath"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("s3 that points to gzip uri") {
@@ -137,7 +137,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"s3://$filePath"
         val expectedPath = s"/vsigzip//vsis3/$filePath"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("s3 that points to gzip with uri") {
@@ -145,7 +145,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"s3://$filePath!$fileName"
         val expectedPath = s"/vsigzip//vsis3/$filePath/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("gzip+s3 uri") {
@@ -153,7 +153,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"gzip+s3://$path"
         val expectedPath = s"/vsigzip//vsis3/$path"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("gzip+s3 uri with !") {
@@ -161,7 +161,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"gzip+s3://$path!$fileName"
         val expectedPath = s"/vsigzip//vsis3/$path/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("s3 that points to gz uri") {
@@ -169,7 +169,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"s3://$filePath"
         val expectedPath = s"/vsigzip//vsis3/$filePath"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("s3 that points to gz with uri") {
@@ -177,7 +177,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"s3://$filePath!$fileName"
         val expectedPath = s"/vsigzip//vsis3/$filePath/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("gzip+s3 uri for a .gz ext") {
@@ -185,7 +185,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"gzip+s3://$path"
         val expectedPath = s"/vsigzip//vsis3/$path"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("gzip+s3 uri with ! for a .gz ext") {
@@ -193,7 +193,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"gzip+s3://$path!$fileName"
         val expectedPath = s"/vsigzip//vsis3/$path/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
     }
 
@@ -203,7 +203,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"hdfs://$filePath"
         val expectedPath = s"/vsihdfs/$uri"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("hdfs that points to tgz uri") {
@@ -211,7 +211,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"hdfs://$filePath"
         val expectedPath = s"/vsitar//vsihdfs/$uri"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("hdfs that points to tgz with ! uri") {
@@ -219,7 +219,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"hdfs://$filePath!$fileName"
         val expectedPath = s"/vsitar//vsihdfs/$filePath/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("zip+hdfs uri") {
@@ -227,7 +227,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"zip+hdfs://$filePath"
         val expectedPath = s"/vsizip//vsihdfs/$uri"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("zip+hdfs with ! uri") {
@@ -235,7 +235,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"zip+hdfs://$filePath!$fileName"
         val expectedPath = s"/vsizip//vsihdfs/$filePath/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
     }
 
@@ -245,7 +245,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"gs://$filePath"
         val expectedPath = s"/vsigs/$filePath"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("Google Cloud Storage that points to tar uri") {
@@ -253,7 +253,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"gs://$filePath"
         val expectedPath = s"/vsitar//vsigs/$filePath"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("Google Cloud Storage that points to tar with ! uri") {
@@ -261,7 +261,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"gs://$filePath!$fileName"
         val expectedPath = s"/vsitar//vsigs/$filePath/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("tar+gs uri") {
@@ -269,7 +269,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"tar+gs://$filePath"
         val expectedPath = s"/vsitar//vsigs/$filePath"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("tar+gs with ! uri") {
@@ -277,7 +277,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"tar+gs://$filePath!$fileName"
         val expectedPath = s"/vsitar//vsigs/$filePath/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
     }
 
@@ -286,28 +286,28 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = "wasb://test-files@myaccount.blah.core.net/nlcd/data/tiff-0.tiff"
         val expectedPath = "/vsiaz/test-files/nlcd/data/tiff-0.tiff"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("Azure that points to kmz uri") {
         val uri = "wasb://test-files@myaccount.blah.core.net/nlcd/data/info.kmz"
         val expectedPath = "/vsizip//vsiaz/test-files/nlcd/data/info.kmz"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("Azure that points to kmz with ! uri") {
         val uri = s"wasb://test-files@myaccount.blah.core.net/nlcd/data/info.kmz!$fileName"
         val expectedPath = s"/vsizip//vsiaz/test-files/nlcd/data/info.kmz/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("wasb+zip uri") {
         val uri = "zip+wasb://test-files@myaccount.blah.core.net/nlcd/data/info.zip"
         val expectedPath = "/vsizip//vsiaz/test-files/nlcd/data/info.zip"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("wasb+zip with ! uri") {
@@ -315,7 +315,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"$path!$fileName"
         val expectedPath = s"/vsizip//vsiaz/test-files/nlcd/data/info.zip/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
     }
 
@@ -325,7 +325,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = filePath
         val expectedPath = filePath
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("relative path that points to zip uri") {
@@ -333,7 +333,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = filePath
         val expectedPath = s"/vsizip/$filePath"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
 
       it("relative path that points to zip with ! uri") {
@@ -341,7 +341,7 @@ class VSIPathSpec extends FunSpec with Matchers {
         val uri = s"$filePath!$fileName"
         val expectedPath = s"/vsizip/$filePath/$fileName"
 
-        VSIPath(uri).vsiPath should be (expectedPath)
+        GDALDataPath(uri).formattedPath should be (expectedPath)
       }
     }
   }
@@ -352,7 +352,7 @@ class VSIPathSpec extends FunSpec with Matchers {
       val uri = s"s3://$filePath"
       val expectedPath = s"/vsis3/$filePath"
 
-      VSIPath(uri).vsiPath should be (expectedPath)
+      GDALDataPath(uri).formattedPath should be (expectedPath)
     }
 
     it("should parse a targeted compressed file with a differenct delimiter") {
@@ -360,7 +360,7 @@ class VSIPathSpec extends FunSpec with Matchers {
       val uri = s"s3://$filePath/$fileName"
       val expectedPath = s"/vsis3/$filePath/$fileName"
 
-      VSIPath(uri, "/").vsiPath should be (expectedPath)
+      GDALDataPath(uri, "/").formattedPath should be (expectedPath)
     }
   }
 }
